backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# HTTP client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Code quality
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
